requirements.txt
altgraph==0.17.4
anticaptchaofficial==1.0.62
attrs==24.3.0
beautifulsoup4==4.13.2
blinker==1.9.0
bs4==0.0.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cx_Freeze==7.2.10
cx_Logging==3.2.1
et_xmlfile==2.0.0
fdb==2.0.2
firebird-base==1.8.0
firebird-driver==1.10.9
firebirdsql==1.3.2
Flask==3.1.0
Flask-Cors==5.0.0
future==1.0.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lief==0.16.2
MarkupSafe==3.0.2
Nuitka==2.6.4
numpy==2.2.2
openpyxl==3.1.5
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pefile==2023.2.7
pillow==11.1.0
protobuf==5.29.3
psycopg2==2.9.10
py==1.11.0
pycparser==2.22
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.0
pyodbc==5.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PySocks==1.7.1
python-anticaptcha==1.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
pywin32-ctypes==0.2.3
requests==2.32.3
selenium==4.28.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.37
sqlalchemy-firebird==2.1
trio==0.28.0
trio-websocket==0.11.1
ttkthemes==3.2.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zstandard==0.23.0
